Add files via upload
UPI CUSTOMER CHURN PREDICTION
The concern dataset is about a UPI payments platform where information about referral, permissions given by the customer to the platform, rewards offered by the platform to the Customer, City, Amount of the payment etc is provided. This is a classification problem where whether the customer will churn out of the platform or not is to be predicted.